+++ blueprints/ember-flexberry/index.js
@@ -69,6 +69,7 @@ module.exports = {
       '        storeDeprecationMessages: false,\n' +
       '        storePromiseErrors: true,\n' +
       '        showPromiseErrors: true,\n' +
+      '        errorMessageFilterActive: true,\n' +
       '      },\n\n' +
       '      // Options for Perforator service that can be used to calculate performance of components rendering.\n' +
       '      perf: {\n' +