+++ tests/dummy/config/environment.js
@@ -51,6 +51,7 @@ module.exports = function(environment) {
         storeDeprecationMessages: true,
         storePromiseErrors: true,
         showPromiseErrors: true,
+        errorMessageFilterActive: true,
       },
 
       perf: {