+++ addon/components/object-list-view.js
@@ -1270,8 +1270,6 @@ export default FlexberryBaseComponent.extend(
           let renderedRowIndex = this.get('_renderedRowIndex') + 1;
 
           if (renderedRowIndex >= contentLength) {
-            this._restoreSelectedRecords();
-
             // The last menu needs will be up.
             this.$('.object-list-view-menu:last .ui.dropdown').addClass('bottom');
             this.$('.object-list-view-menu > .ui.dropdown').dropdown();
@@ -1314,8 +1312,6 @@ export default FlexberryBaseComponent.extend(
         }
       }
     } else {
-      this._restoreSelectedRecords();
-
       if (this.get('allowColumnResize')) {
         this._reinitResizablePlugin();
       } else {
@@ -1336,6 +1332,17 @@ export default FlexberryBaseComponent.extend(
 
       this._fixedTableHead($currentTable);
     }
+  },
+
+  /**
+    Restores the state of selected rows after updating the list.
+
+    See [EmberJS API](https://api.emberjs.com/).
+
+    @method didUpdateAttrs
+  */
+  didUpdateAttrs() {
+    Ember.run.scheduleOnce('afterRender', this, this._restoreSelectedRecords);
   },
 
   /**