+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 * The `applicationLogModelName` property for the `log` service.
+### Changed
+* `log` service:
+    * Disable transition aborted errors save.
 
 ### Changed
 * Properties initialization of `log` service moved to the service's init hook (instead of reopening service in app).