+++ CHANGELOG.md
@@ -3,7 +3,16 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* In actions for customize filters for components of lists, a parameter has been added with a description of the attribute for which the action is called.
+* The `applicationLogModelName` property for the `log` service.
+
+### Changed
+* Properties initialization of `log` service moved to the service's init hook (instead of reopening service in app).
+
+### Fixed
 * `flexberry-objectlistview` component:
+    * Fix menu opening direction in line.
     * Add row button at mobile version.
 
 ## [2.4.0-beta.0] - 2020-03-18