+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+* `log` service:
+    * Disable save to log transition aborted errors.
 
 ## [2.4.0-beta.1] - 2020-03-23
 ### Added